+++ booking/src/main/java/com/appointment/booking/service/impl/ProviderServiceImpl.java
@@ -7,6 +7,8 @@ import com.appointment.booking.repository.ProviderRepository;
 import com.appointment.booking.service.ProviderService;
 import com.appointment.booking.service.UserService;
 import lombok.RequiredArgsConstructor;
+import com.appointment.booking.model.UserRole;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -16,8 +18,9 @@ import java.util.Optional;
 @RequiredArgsConstructor
 public class ProviderServiceImpl implements ProviderService {
 
+    @Autowired
     private final ProviderRepository providerRepository;
-
+    @Autowired
     private final UserService userService;
 
     @Override
@@ -31,11 +34,18 @@ public class ProviderServiceImpl implements ProviderService {
                 .orElseThrow(() -> new RuntimeException("Provider not found"));
         return providerRepository.findById(id);
     }
+    public Optional<Provider> findByUser(User user) {
+        return providerRepository.findByUser(user);
+    }
 
     @Override
-    public Provider updateProvider(Long id, ProviderRequest req) {
+    public Provider updateProvider(Long id, ProviderRequest req, User currentUser) {
         Provider provider = providerRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Provider not found"));
+        if (!provider.getUser().getId().equals(currentUser.getId()) 
+            && currentUser.getRole() != UserRole.ADMIN) {
+        throw new RuntimeException("Not authorized");
+        }
 
         // Optional: update linked user if userId provided
         if (req.getUserId() != null) {
@@ -51,9 +61,13 @@ public class ProviderServiceImpl implements ProviderService {
         return providerRepository.save(provider);
     }
     @Override
-    public void deleteProvider(Long id) {
+    public void deleteProvider(Long id, User currentUser) {
         Provider provider = providerRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Provider not found"));
+        if (!provider.getUser().getId().equals(currentUser.getId()) 
+            && currentUser.getRole() != UserRole.ADMIN) {
+         throw new RuntimeException("You are not authorized to delete this provider");
+            }
         providerRepository.delete(provider);
     }
